Test managers

+++ core/src/test/java/metamodels/Role_.java
@@ -9,7 +9,7 @@ import utils.AttributeMock;
 
 @StaticMetamodel(Role.class)
 public class Role_ {
-    public static SingularAttribute<Role, Integer> id = AttributeMock.of("id", Integer.class);
+    public static SingularAttribute<Role, Long> id = AttributeMock.of("id", Long.class);
     public static SingularAttribute<Role, String> name = AttributeMock.of("name", String.class);
     public static ListAttribute<Role, Preferences> permissions = AttributeMock.ofList("permissions", Preferences.class);
 }
